Extract non-DEM texture image processing
